Fixing Issues 12 - 15
This revision contains bug fixes for issues # 12 -15.  Most of these
issues are related to the filtering of Types and Components.  In the
past, the IDM plugins filtered out non-COBie assets, but the IDM plugins
are no longer supported.

Also, the COBieExcelTemplate.xlsx was revised to remove blank lines as
reported in Issue #13.